Make PipelineThinker a subclass of Single
Was about to refactor in a way that the Single objective
thinker did not have any methods
